update: 实现 entropy，并跑出结果
